refactor(component-meta): rename APIs
