+++ packages/component-meta/package.json
@@ -15,7 +15,7 @@
 	"dependencies": {
 		"@volar/typescript": "~1.10.9",
 		"@vue/language-core": "1.8.22",
-		"typesafe-path": "^0.2.2",
+		"path-browserify": "^1.0.1",
 		"vue-component-type-helpers": "1.8.22"
 	},
 	"peerDependencies": {
@@ -27,6 +27,7 @@
 		}
 	},
 	"devDependencies": {
-		"@types/node": "latest"
+		"@types/node": "latest",
+		"@types/path-browserify": "latest"
 	}
 }
