+++ packages/component-meta/src/types.ts
@@ -1,5 +1,7 @@
 import type * as ts from 'typescript/lib/tsserverlibrary';
 
+export type Checker = ReturnType<typeof import('./base')['baseCreate']>;
+
 export interface Declaration {
 	file: string;
 	range: [number, number];
